restrict subclass conversion to ndarray
To provide backwards compability, the ndarray
converter is allowed to convert subclasses of
ndarray that are not handled by other converters.
This subclass conversion results in a AsdfConversionWarning
as a future ASDF version will not support this behavior
(as converters require explicit definition of what
types are handled).
To aid in testing, a configuration option is added
`convert_unknown_ndarray_subclasses` that is True by
default but can be turned off to not allow conversion
of subclasses (to more closely match the behavior
of a future version of ASDF).