add Converter.change_type, add Reference Converter
allows converters to convert objects to different
types during serialization to allow other converters
to handle the actual serialization of the object.

This allows Reference to be handled by a Converter
and provides a means by which downstream packages
(like asdf-astropy and stdatamodels) can convert
ndarray subclasses to ndarray during serialization.